// packages/db/src/repositories/jobsRepository.d.ts
import type { DrizzleD1Database } from 'drizzle-orm/d1';
import type { JobModel } from '../schema';
export declare class JobsRepository {
    private db;
    constructor(db: DrizzleD1Database<any>);
    createJob(job: JobModel): Promise<void>;
    getJobById(id: string): Promise<{
        id: string;
        company: string;
        title: string;
        url: string;
        score: number;
        status: string;
        location: string;
        modality: string;
        easyApply: boolean;
        language: string;
        profile: string;
        createdAt: Date;
        updatedAt: Date;
        applyResult: string | null;
        postedAt: string | null;
        description: string | null;
    } | undefined>;
    getJobByUrl(url: string): Promise<{
        id: string;
        company: string;
        title: string;
        url: string;
        score: number;
        status: string;
        location: string;
        modality: string;
        easyApply: boolean;
        language: string;
        profile: string;
        createdAt: Date;
        updatedAt: Date;
        applyResult: string | null;
        postedAt: string | null;
        description: string | null;
    } | undefined>;
    getJobsByStatus(status: string): Promise<{
        id: string;
        company: string;
        title: string;
        url: string;
        score: number;
        status: string;
        location: string;
        modality: string;
        easyApply: boolean;
        language: string;
        profile: string;
        createdAt: Date;
        updatedAt: Date;
        applyResult: string | null;
        postedAt: string | null;
        description: string | null;
    }[]>;
    getAllJobs(): Promise<{
        id: string;
        company: string;
        title: string;
        url: string;
        score: number;
        status: string;
        location: string;
        modality: string;
        easyApply: boolean;
        language: string;
        profile: string;
        createdAt: Date;
        updatedAt: Date;
        applyResult: string | null;
        postedAt: string | null;
        description: string | null;
    }[]>;
    upsertJob(job: JobModel): Promise<void>;
    updateJobStatus(id: string, status: string, applyResult?: string): Promise<void>;
}